+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/TigerDirector.java
@@ -4,6 +4,7 @@
 
 package de.gematik.test.tiger.lib;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 import de.gematik.rbellogger.RbelOptions;
 import de.gematik.rbellogger.util.RbelAnsiColors;
@@ -22,8 +23,6 @@ import de.gematik.test.tiger.testenvmgr.env.TigerStatusUpdate;
 import de.gematik.test.tiger.testenvmgr.util.TigerEnvironmentStartupException;
 import de.gematik.test.tiger.testenvmgr.util.TigerTestEnvException;
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.ArrayList;
@@ -31,11 +30,12 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 import java.util.regex.Pattern;
-import javax.annotation.PreDestroy;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import net.serenitybdd.rest.SerenityRest;
 import org.apache.commons.io.IOUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.assertj.core.api.Fail;
 import org.awaitility.core.ConditionTimeoutException;
 import org.springframework.boot.Banner.Mode;
 import org.springframework.boot.WebApplicationType;
@@ -93,7 +93,9 @@ public class TigerDirector {
         Runtime.getRuntime().addShutdownHook(new Thread(() -> {
             try {
                 if (getLibConfig().isActivateWorkflowUi() && !tigerTestEnvMgr.isUserAcknowledgedShutdown()) {
-                    System.out.println(Ansi.colorize("TGR Workflow UI is active, please press quit in browser window...", RbelAnsiColors.GREEN_BOLD));
+                    System.out.println(
+                        Ansi.colorize("TGR Workflow UI is active, please press quit in browser window...",
+                            RbelAnsiColors.GREEN_BOLD));
                     if (tigerTestEnvMgr != null) {
                         tigerTestEnvMgr.receiveTestEnvUpdate(TigerStatusUpdate.builder()
                             .bannerMessage("Test run finished, press QUIT")
@@ -180,10 +182,14 @@ public class TigerDirector {
     private static synchronized void startWorkflowUi() {
         if (libConfig.activateWorkflowUi) {
             log.info("\n" + Banner.toBannerStr("STARTING WORKFLOW UI ...", RbelAnsiColors.BLUE_BOLD.toString()));
-            TigerTestEnvMgr.openWorkflowUiInBrowser(TigerGlobalConfiguration.readIntegerOptional("tiger.internal.testenvmgr.port").orElseThrow(() -> new TigerEnvironmentStartupException("No free port for test environment manager reserved!")).toString());
+            TigerTestEnvMgr.openWorkflowUiInBrowser(
+                TigerGlobalConfiguration.readIntegerOptional("tiger.internal.testenvmgr.port").orElseThrow(
+                        () -> new TigerEnvironmentStartupException("No free port for test environment manager reserved!"))
+                    .toString());
             log.info("Waiting for workflow Ui to fetch status...");
             try {
-                await().atMost(Duration.ofSeconds(10)).pollInterval(Duration.ofSeconds(1)).until(() -> tigerTestEnvMgr.isWorkflowUiSentFetch());
+                await().atMost(Duration.ofSeconds(10)).pollInterval(Duration.ofSeconds(1))
+                    .until(() -> tigerTestEnvMgr.isWorkflowUiSentFetch());
             } catch (ConditionTimeoutException cte) {
                 libConfig.activateWorkflowUi = false;
                 throw new TigerTestEnvException("No feedback from workflow Ui, aborting!", cte);
@@ -205,7 +211,9 @@ public class TigerDirector {
                 log.info(Ansi.colorize("SKIPPING TIGER PROXY settings as System Property is set already...",
                     RbelAnsiColors.RED_BOLD));
             } else {
-                log.info(Ansi.colorize("SETTING TIGER PROXY http://localhost:" + tigerTestEnvMgr.getLocalTigerProxy().getProxyPort() + "...", RbelAnsiColors.BLUE_BOLD));
+                log.info(Ansi.colorize(
+                    "SETTING TIGER PROXY http://localhost:" + tigerTestEnvMgr.getLocalTigerProxy().getProxyPort()
+                        + "...", RbelAnsiColors.BLUE_BOLD));
                 System.setProperty("http.proxyHost", "localhost");
                 System.setProperty("http.proxyPort", "" + tigerTestEnvMgr.getLocalTigerProxy().getProxyPort());
                 System.setProperty("http.nonProxyHosts", "localhost|127.0.0.1");
@@ -215,7 +223,8 @@ public class TigerDirector {
                 SerenityRestUtils.setupSerenityRest(tigerTestEnvMgr.getLocalTigerProxy().getProxyPort());
             }
         } else {
-            log.info(Ansi.colorize("SKIPPING TIGER PROXY settings as localProxyActive==false...", RbelAnsiColors.RED_BOLD));
+            log.info(
+                Ansi.colorize("SKIPPING TIGER PROXY settings as localProxyActive==false...", RbelAnsiColors.RED_BOLD));
         }
 
         // TODO TGR-295 DO NOT DELETE!
@@ -261,7 +270,8 @@ public class TigerDirector {
         }
     }
 
-    private final static Pattern showSteps = Pattern.compile(".*TGR (zeige|show) ([\\w|ü|ß]*) (Banner|banner|text|Text) \"(.*)\"");//NOSONAR
+    private final static Pattern showSteps = Pattern.compile(
+        ".*TGR (zeige|show) ([\\w|ü|ß]*) (Banner|banner|text|Text) \"(.*)\"");//NOSONAR
 
     private static void assertThatTigerIsInitialized() {
         if (!initialized) {
@@ -316,18 +326,53 @@ public class TigerDirector {
     }
 
     public static void pauseExecution() {
+        pauseExecution("");
+    }
+
+    public static void pauseExecution(String message) {
+        String defaultMessage = "Test execution paused, click to continue";
+        if (StringUtils.isBlank(message)) {
+            message = defaultMessage;
+        }
+
         if (getLibConfig().isActivateWorkflowUi()) {
             tigerTestEnvMgr.receiveTestEnvUpdate(TigerStatusUpdate.builder()
-                .bannerMessage("Test execution paused, click to continue")
+                .bannerMessage(message)
                 .bannerColor("green")
                 .bannerType(BannerType.STEP_WAIT)
                 .build());
             await().pollInterval(1, TimeUnit.SECONDS)
                 .atMost(5, TimeUnit.HOURS)
                 .until(() -> tigerTestEnvMgr.isUserAcknowledgedContinueTestRun());
-            tigerTestEnvMgr.resetUserAcknowledgedContinueTestRun();
+            tigerTestEnvMgr.resetUserInput();
         } else {
-            tigerTestEnvMgr.waitForConsoleInput("next");
+            // TGR-585
+            log.warn(String.format("The step 'TGR pause test run execution with message \"%s\"' is not supported outside the Workflow UI. Please check the manual for more information.", message));
+
+        }
+    }
+
+    public static void pauseExecutionAndFailIfDesired(String message, String errorMessage) {
+        if (getLibConfig().isActivateWorkflowUi()) {
+            tigerTestEnvMgr.receiveTestEnvUpdate(TigerStatusUpdate.builder().bannerMessage(
+                    message)
+                .bannerColor("black")
+                .bannerType(BannerType.FAIL_PASS)
+                .build());
+            await().pollInterval(1, TimeUnit.SECONDS)
+                .atMost(5, TimeUnit.HOURS)
+                .until(() -> tigerTestEnvMgr.isUserAcknowledgedContinueTestRun()
+                    || tigerTestEnvMgr.isUserAcknowledgedFailingTestRun());
+            if (tigerTestEnvMgr.isUserAcknowledgedFailingTestRun()) {
+                tigerTestEnvMgr.resetUserInput();
+                Fail.fail(errorMessage);
+            } else {
+                tigerTestEnvMgr.resetUserInput();
+            }
+        } else {
+            // TGR-585
+            log.warn(String.format(
+                "The step 'TGR pause test run execution with message \"%s\" and message in case of error \"%s\"' is not supported outside the Workflow UI. Please check the manual for more information.", message, errorMessage));
         }
     }
 }