+++ tiger-testenv-mgr/src/frontend/src/types/BannerType.ts
@@ -2,6 +2,7 @@ enum DataType {
   MESSAGE = "MESSAGE",
   STEP_WAIT = "STEP_WAIT",
   TESTRUN_ENDED = "TESTRUN_ENDED",
+  FAIL_PASS = "FAIL_PASS"
 }
 
 export default DataType;